Actualizado

+++ src/java/com/tartanga/grupo4/transfers/Transfers.java
@@ -49,7 +49,7 @@ public class Transfers implements Serializable {
     @Temporal(TemporalType.DATE)
     private Date transferDate;
     
-    private Integer Amount;
+    private Double Amount;
     
     @Enumerated(EnumType.STRING)
     private Currency currency;
@@ -65,10 +65,10 @@ public class Transfers implements Serializable {
         this.sender="";
         this.reciever="";
         this.transferDate=null;
-        this.Amount=0;
+        this.Amount=0.0;
         this.currency=Currency.EURO;
     }
-    public Transfers(Integer transferId,String sender,String reciever,Date transferDate,Integer Amount,Currency currency){
+    public Transfers(Integer transferId,String sender,String reciever,Date transferDate,Double Amount,Currency currency){
         this.transferId=transferId;
         this.sender=sender;
         this.reciever=reciever;
@@ -109,11 +109,11 @@ public class Transfers implements Serializable {
         this.transferDate = transferDate;
     }
 
-    public Integer getAmount() {
+    public Double getAmount() {
         return Amount;
     }
 
-    public void setAmount(Integer Amount) {
+    public void setAmount(Double Amount) {
         this.Amount = Amount;
     }
